requirements.txt
Flask
werkzeug
google-generativeai
python-dotenv
sentence-transformers
faiss-cpu
pillow
open-ai
